Cypress tests are working
This is the last CL in the chain other than review comments
