+++ package.json
@@ -30,19 +30,19 @@
   },
   "devDependencies": {
     "@bazel/bazelisk": "^1.9.0",
-    "cypress": "7.3.0",
     "@grpc/grpc-js": "^1.7.3",
-    "google-protobuf": "^3.19.1",
-    "grpc-web": "^1.4.2",
-    "jasmine": "^3.7.0",
-    "jasmine-core": "^3.7.1",
-    "typescript": "^4.9.3",
     "@rollup/plugin-commonjs": "^23.0.2",
     "@rollup/plugin-node-resolve": "^15.0.1",
-    "rollup": "^3.3.0",
-    "http-server": "^14.1.1",
     "@types/google-protobuf": "^3.15.5",
     "@types/jasmine": "^3.7.7",
-    "@types/node": "^18.11.9"
+    "@types/node": "^18.11.9",
+    "cypress": "10.8.0",
+    "google-protobuf": "^3.19.1",
+    "grpc-web": "^1.4.2",
+    "http-server": "^14.1.1",
+    "jasmine": "^3.7.0",
+    "jasmine-core": "^3.7.1",
+    "rollup": "^3.3.0",
+    "typescript": "^4.9.3"
   }
 }
