--- test/web/cypress.json
@@ -1,1 +1,0 @@
-{}
